Up the apache timeout
For a while we have been having trouble cloning nova from git.o.o,
particularly during busy times of the day or if we have a job that
needs to clone it multiple times across workers (turbo-hipster for
example).

After investigating the tcpdumps it appears to be because the
origin sends a RST during the git pack headers which regularly
takes more than 60 seconds now.

Change-Id: Id1078dfaed2a8e17ec086e13b53df51cb59d480c