Ensure cgit dir is created after parent directory
In the default case, the cgit directory is /var/www/cgit. The parent
directory, /var/www, is created by the apache package. If the package
has not yet been installed, the file resource will fail with:

 Error: Cannot create /var/www/cgit; parent directory /var/www does not exist

This package fixes the issue by adding a require relationship between
the apache package and the cgit directory. The Package['httpd']
resource is declared in the apache class, which is included at the top
of this manifest.

Change-Id: Ifba16ee06ed485b539c576e53a68450fad6706d3