Fix selinux tests
If things are working correctly, then there may only be one port number
for a given port type, so it may not be part of a comma-separated list.
This patch relaxes the tests to allow a single port number instead of
strictly a comma-separated list in the output of `semanage port -l`.

Change-Id: Iaa9c8cda7a2eae904eb8f25cfa33be249b2b4cab