Fix beaker
Since the beaker jobs are being run on xenial, we need a special nodeset
for it, otherwise beaker gives an error:

 beaker-hostgenerator was not able to use this value as input.
 Exiting with an Error.

We also want to install puppet from the Ubuntu repos rather than from
puppetlabs, since puppetlabs doesn't support puppet 3 for Xenial. For
centos we can keep the install process the same.

Finally, since the epel repo is now disabled by default on nodepool
nodes, make sure it's enabled in the package resource.

Change-Id: Ifd2244ae9dd212b2475f9cd6adb994bc058a4769