Add order to cgit package
The httpd module uses the file resource with the purge option to clear
out the httpd conf directory. On puppet 4, the resource ordering
algorithm changed such that the directory purge happens before the cgit
RPM adds a config file there, which means on the next puppet run it
purges it again and bounces the service again. This cause the
idempotency test in beaker to fail.

This patch adds an ordering parameter to ensure that the cgit package is
installed before the httpd class runs so that it doesn't have to clear
out the httpd config directory twice. Since puppet 4 more or less tries
to order resources in the order they appear, also swap the package
resourcs and httpd class just to make it clear what order things should
be in.

Change-Id: I813f6e9f82d3b44b1d38fb5773c5bd6160f58b78